+++ mod-record-specifications-server/src/main/java/org/folio/rspec/service/mapper/FieldIndicatorMapper.java
@@ -5,11 +5,13 @@ import org.folio.rspec.domain.dto.FieldIndicatorChangeDto;
 import org.folio.rspec.domain.dto.FieldIndicatorDto;
 import org.folio.rspec.domain.entity.Indicator;
 import org.mapstruct.AfterMapping;
+import org.mapstruct.BeanMapping;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.MappingConstants;
 import org.mapstruct.MappingTarget;
 import org.mapstruct.Named;
+import org.mapstruct.NullValuePropertyMappingStrategy;
 import org.mapstruct.ReportingPolicy;
 
 @Mapper(unmappedTargetPolicy = ReportingPolicy.WARN, componentModel = MappingConstants.ComponentModel.SPRING,
@@ -34,6 +36,13 @@ public interface FieldIndicatorMapper {
   @Mapping(target = "metadata", ignore = true)
   @Mapping(target = "id", ignore = true)
   Indicator toEntity(FieldIndicatorChangeDto createDto);
+
+  @Mapping(target = "id", ignore = true)
+  @Mapping(target = "field", ignore = true)
+  @Mapping(target = "codes", ignore = true)
+  @Mapping(target = "metadata", ignore = true)
+  @BeanMapping(nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.SET_TO_NULL)
+  void update(@MappingTarget Indicator indicator, FieldIndicatorChangeDto changeDto);
 
   @AfterMapping
   default void resetCollections(@MappingTarget FieldIndicatorDto target) {